blast-embedding/requirements.txt
numpy==1.21.0
scikit-learn==0.24.2
transformers==4.10.0
gensim==4.0.1
biopython==1.79
torch==1.9.0